requirements.txt
pystray
pygetwindow
pillow
pywin32